test(relay): update test cases
