+++ packages/next-swc/crates/relay/Cargo.toml
@@ -16,3 +16,7 @@ serde = "1"
 serde_json = "1"
 regex = "1.5"
 once_cell = "1.8.0"
+
+[dev-dependencies]
+swc_ecma_transforms_testing = "0.87.0"
+testing = "0.20.0"
